Add new Hydra Builder Interface and implementations
