+++ src/main/java/eu/nampi/backend/model/hydra/InterfaceHydraBuilder.java
@@ -1,11 +1,7 @@
 package eu.nampi.backend.model.hydra;
 
-import org.apache.jena.query.Query;
-
 public interface InterfaceHydraBuilder {
 
-  public Query build();
-
-  public String buildString();
+  public String buildHydra();
 
 }
