+++ src/main/java/eu/nampi/backend/model/hydra/InterfaceHydraBuilderOld.java
@@ -1,0 +1,11 @@
+package eu.nampi.backend.model.hydra;
+
+import org.apache.jena.query.Query;
+
+public interface InterfaceHydraBuilderOld {
+
+  public Query build();
+
+  public String buildString();
+
+}
