+++ src/main/java/eu/nampi/backend/repository/AuthorRepository.java
@@ -3,6 +3,7 @@ package eu.nampi.backend.repository;
 import java.util.Optional;
 import java.util.UUID;
 import java.util.concurrent.atomic.AtomicReference;
+
 import org.apache.jena.arq.querybuilder.SelectBuilder;
 import org.apache.jena.arq.querybuilder.UpdateBuilder;
 import org.apache.jena.arq.querybuilder.WhereBuilder;
@@ -17,10 +18,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Repository;
+
 import eu.nampi.backend.model.Author;
 import eu.nampi.backend.model.QueryParameters;
-import eu.nampi.backend.model.hydra.HydraCollectionBuilder;
-import eu.nampi.backend.model.hydra.HydraSingleBuilder;
+import eu.nampi.backend.model.hydra.HydraCollectionBuilderOld;
+import eu.nampi.backend.model.hydra.HydraSingleBuilderOld;
 import eu.nampi.backend.service.JenaService;
 import eu.nampi.backend.vocabulary.Core;
 import eu.nampi.backend.vocabulary.Doc;
@@ -35,13 +37,11 @@ public class AuthorRepository extends AbstractHydraRepository {
   private JenaService jenaService;
 
   public Model findAll(QueryParameters params) {
-    HydraCollectionBuilder hydra =
-        new HydraCollectionBuilder(params, Core.author, Doc.authorOrderByVar);
+    HydraCollectionBuilderOld hydra = new HydraCollectionBuilderOld(params, Core.author, Doc.authorOrderByVar);
     return construct(hydra);
   }
 
-  @Cacheable(
-      key = "{#lang, #params.limit, #params.offset, #params.orderByClauses, #params.type, #params.text}")
+  @Cacheable(key = "{#lang, #params.limit, #params.offset, #params.orderByClauses, #params.type, #params.text}")
   public String findAll(QueryParameters params, Lang lang) {
     Model model = findAll(params);
     return serialize(model, lang, ResourceFactory.createResource(params.getBaseUrl()));
@@ -50,7 +50,7 @@ public class AuthorRepository extends AbstractHydraRepository {
   @Cacheable(key = "{#lang, #id}")
   public String findOne(Lang lang, UUID id) {
     String uri = individualsUri(Core.author, id);
-    HydraSingleBuilder builder = new HydraSingleBuilder(uri, Core.author);
+    HydraSingleBuilderOld builder = new HydraSingleBuilderOld(uri, Core.author);
     Model model = construct(builder);
     return serialize(model, lang, ResourceFactory.createResource(uri));
   }
@@ -60,8 +60,8 @@ public class AuthorRepository extends AbstractHydraRepository {
     SelectBuilder selectBuilder = new SelectBuilder();
     String iri = individualsUri(Core.author, rdfId);
     try {
-      selectBuilder.addValueVar("?l").addWhere("?a", RDF.type, Core.author)
-          .addWhere("?a", RDFS.label, "?l").addFilter("?a = <" + iri + ">");
+      selectBuilder.addValueVar("?l").addWhere("?a", RDF.type, Core.author).addWhere("?a", RDFS.label, "?l")
+          .addFilter("?a = <" + iri + ">");
     } catch (ParseException e) {
       log.warn(e.getMessage());
     }
@@ -75,8 +75,8 @@ public class AuthorRepository extends AbstractHydraRepository {
   public Author addOne(UUID rdfId, String label) {
     String iri = individualsUri(Core.author, rdfId);
     Resource authorRes = ResourceFactory.createResource(iri);
-    UpdateBuilder updateBuilder = new UpdateBuilder().addInsert(authorRes, RDF.type, Core.author)
-        .addInsert(authorRes, RDFS.label, label);
+    UpdateBuilder updateBuilder = new UpdateBuilder().addInsert(authorRes, RDF.type, Core.author).addInsert(authorRes,
+        RDFS.label, label);
     jenaService.update(updateBuilder);
     return new Author(iri, rdfId, label);
   }
@@ -85,8 +85,8 @@ public class AuthorRepository extends AbstractHydraRepository {
     UpdateBuilder updateBuilder = new UpdateBuilder();
     try {
       updateBuilder.addDelete("?a", RDFS.label, "?l").addInsert("?a", RDFS.label, newLabel)
-          .addWhere(new WhereBuilder().addWhere("?a", RDF.type, Core.author)
-              .addWhere("?a", RDFS.label, "?l").addFilter("?a = <" + author.getIri() + ">"));
+          .addWhere(new WhereBuilder().addWhere("?a", RDF.type, Core.author).addWhere("?a", RDFS.label, "?l")
+              .addFilter("?a = <" + author.getIri() + ">"));
     } catch (ParseException e) {
       log.error(e.getMessage());
     }
