+++ src/main/java/eu/nampi/backend/repository/PlaceRepository.java
@@ -1,16 +1,18 @@
 package eu.nampi.backend.repository;
 
 import java.util.UUID;
+
 import org.apache.jena.rdf.model.Model;
 import org.apache.jena.rdf.model.ResourceFactory;
 import org.apache.jena.riot.Lang;
 import org.springframework.cache.annotation.CacheConfig;
 import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Repository;
+
 import eu.nampi.backend.model.QueryParameters;
-import eu.nampi.backend.model.hydra.AbstractHydraBuilder;
-import eu.nampi.backend.model.hydra.HydraCollectionBuilder;
-import eu.nampi.backend.model.hydra.HydraSingleBuilder;
+import eu.nampi.backend.model.hydra.AbstractHydraBuilderOld;
+import eu.nampi.backend.model.hydra.HydraCollectionBuilderOld;
+import eu.nampi.backend.model.hydra.HydraSingleBuilderOld;
 import eu.nampi.backend.vocabulary.Core;
 import eu.nampi.backend.vocabulary.Doc;
 import eu.nampi.backend.vocabulary.SchemaOrg;
@@ -20,14 +22,12 @@ import eu.nampi.backend.vocabulary.SchemaOrg;
 public class PlaceRepository extends AbstractHydraRepository {
 
   public Model findAll(QueryParameters params) {
-    HydraCollectionBuilder hydra =
-        new HydraCollectionBuilder(params, Core.place, Doc.placeOrderByVar);
+    HydraCollectionBuilderOld hydra = new HydraCollectionBuilderOld(params, Core.place, Doc.placeOrderByVar);
     addData(hydra);
     return construct(hydra);
   }
 
-  @Cacheable(
-      key = "{#lang, #params.limit, #params.offset, #params.orderByClauses, #params.type, #params.text}")
+  @Cacheable(key = "{#lang, #params.limit, #params.offset, #params.orderByClauses, #params.type, #params.text}")
   public String findAll(QueryParameters params, Lang lang) {
     Model model = findAll(params);
     return serialize(model, lang, ResourceFactory.createResource(params.getBaseUrl()));
@@ -36,13 +36,13 @@ public class PlaceRepository extends AbstractHydraRepository {
   @Cacheable(key = "{#lang, #id}")
   public String findOne(Lang lang, UUID id) {
     String uri = individualsUri(Core.place, id);
-    HydraSingleBuilder builder = new HydraSingleBuilder(uri, Core.place);
+    HydraSingleBuilderOld builder = new HydraSingleBuilderOld(uri, Core.place);
     addData(builder);
     Model model = construct(builder);
     return serialize(model, lang, ResourceFactory.createResource(uri));
   }
 
-  private void addData(AbstractHydraBuilder<?> builder) {
+  private void addData(AbstractHydraBuilderOld<?> builder) {
     builder.addMainConstruct(SchemaOrg.sameAs, "?sa").addMainOptional(SchemaOrg.sameAs, "?sa");
   }
 }
