+++ src/main/java/eu/nampi/backend/service/JenaService.java
@@ -1,18 +1,20 @@
 package eu.nampi.backend.service;
 
 import java.util.function.Consumer;
-import org.apache.jena.arq.querybuilder.ConstructBuilder;
+
 import org.apache.jena.arq.querybuilder.SelectBuilder;
 import org.apache.jena.arq.querybuilder.UpdateBuilder;
 import org.apache.jena.query.QuerySolution;
 import org.apache.jena.rdf.model.Model;
+
 import eu.nampi.backend.model.hydra.InterfaceHydraBuilder;
+import eu.nampi.backend.model.hydra.InterfaceHydraBuilderOld;
 
 public interface JenaService {
 
-  public Model construct(ConstructBuilder constructBuilder);
+  public Model construct(InterfaceHydraBuilder constructBuilder);
 
-  public Model construct(InterfaceHydraBuilder hydraBuilder);
+  public Model construct(InterfaceHydraBuilderOld hydraBuilder);
 
   public void initInfCache();
